# backend/antigravity/03_api/openapi.snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ATSS Matarat Backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/incidents/{id}/transition": {
      "post": {
        "tags": [
          "SOC"
        ],
        "summary": "Transition Incident",
        "operationId": "transition_incident_incidents__id__transition_post",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncidentTransitionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentTransitionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/incidents/{id}/escalate": {
      "post": {
        "tags": [
          "SOC"
        ],
        "summary": "Escalate Incident",
        "operationId": "escalate_incident_incidents__id__escalate_post",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncidentTransitionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EscalationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EscalationResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "incident_id": {
            "type": "string",
            "title": "Incident Id"
          },
          "ticket_id": {
            "type": "string",
            "title": "Ticket Id"
          },
          "ticket_status": {
            "type": "string",
            "title": "Ticket Status"
          }
        },
        "type": "object",
        "required": [
          "status",
          "incident_id",
          "ticket_id",
          "ticket_status"
        ],
        "title": "EscalationResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IncidentTransitionRequest": {
        "properties": {
          "to_state": {
            "type": "string",
            "title": "To State"
          },
          "triggered_by": {
            "type": "string",
            "title": "Triggered By"
          }
        },
        "type": "object",
        "required": [
          "to_state",
          "triggered_by"
        ],
        "title": "IncidentTransitionRequest"
      },
      "IncidentTransitionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "updated_at_utc": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At Utc",
            "default": "2026-01-09T16:25:06.992158"
          }
        },
        "type": "object",
        "required": [
          "id",
          "state"
        ],
        "title": "IncidentTransitionResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "auth/login"
          }
        }
      }
    }
  }
}